memmodel: support conditional happens-before edges
The current happens-before model is unable to express edges of the
form "if A observes B, then B happens before A". Effectively, when A
and B occur concurrently, the model forks and A both observes and does
not observe B, but this choice is never recorded in the happens-before
graph. As a result, in a program like

T0      T1
a=ld 0  st 0
b=ld 0

There's no way to express that the loads race with the store, but also
must be coherent with each other (that is, if a sees the store, b must
also see the store).

Most weak memory models depend on such conditional edges, so we need
to model them.

Fix this by letting generators return a third "conditional edge"
result. In this case, the model can still chose whether A observes B,
but on the branch where A observes B, it also adds a B -> A edge. In
the above program, this forces the loads to be coherent when combined
with the program order.