benchmany: always checkout before building benchmark
Currently we don't bother checking out the revision if we're using
gover. However, this assumes that the only relevant changes will be to
the toolchain itself. If the benchmark is in the repository being
tested, there may be relevant changes to the benchmark itself (such as
bug fixes) that we want to pick up. Currently we'll miss these changes
and, if we're in randomized mode, this could affect a random subset of
hashes.

Fix this by always checking out the commit hash before doing a build,
regardless of how we're doing the build.