gover: use current Git repo as Go root
Currently gover uses runtime.GOROOT() as the root of the Go tree. As a
result, the save and build commands will operate on the Go tree used
to build gover, even if you're in a different Go tree, which is
extremely confusing.

Fix this by using the Git repository containing the working directory
if it looks like a Go tree and otherwise aborting the save and build
commands.