benchmany: more robust run count tracking
benchmany determines how many runs of a benchmark have completed by
looking for either "PASS", a build failure recorded by benchmany, or
the x/benchmarks output. However, if the benchmark outputs none of
these, benchmany will always think it's run zero iterations of the
benchmark and start from the beginning.

Make this more robust by checking if the output from the command
counts as a run before appending it to the log. If it doesn't, add
"PASS" to the output so it will count.

This means we can also remove the hack for counting x/benchmarks
output, since this is a more general solution.