gover: make run work for arbitrary commands
Currently the run subcommand builds an absolute path to the command to
run assuming it's in the bin directory of the requested tree. This
isn't very useful if you want to run something with a specific GOROOT
that isn't the go command itself.

So, instead, just run the command as specified, but build an
appropriate PATH and GOROOT so that the command will be found on the
PATH if it's in the tree's bin, but doesn't have to be.