benchmany: make benchmarks that depend on the Go tree work
Currently, if a benchmark depends on the Go tree in any way (for
example, if it compiles something), benchmany's optimization to avoid
checking out the Go tree between runs will lead to nonsensical
results. However, simply checking out the right version isn't enough
because the pkg and bin directories will still be out of date. Hence,
what we really want for such benchmarks is the whole tree exactly as
saved by gover. However, currently we only use gover for the build
step and not for running the benchmark.

Make it possible to run benchmarks like this by not only building them
under gover, but also running them under gover. This generalizes the
-gover-save flag into a new -save-tree flag.