gover: fix run for go commands
86a2394 made run work for arbitrary commands, but in the process it
broke running commands from the Go tree bin. In particular, the
command was resolved using the current PATH, even though it was run
with the augmented PATH. As a result, if you tried to run the go
command it would be taken from whatever was in your PATH (even though
it would ultimately run the right compiler by consulting GOROOT). If
there was no go binary in your PATH, it would simply fail.

Fix this by setting gover's PATH before resolving the run command.