rtcheck: add simple path sensitivity
This adds simple path sensitivity to the lockset analysis by tracking
the trivially known values of ssa.Values that affect branches. This
handles the correlated control flow around locks in persistentalloc1
and heapBitsSetType, however, the path sensitivity is currently not
inter-procedural, so it doesn't handle the multi-function correlated
control flow between traceAcquireBuffer and traceReleaseBuffer.

In some cases this can lead to loop unrolling. Currently we bound the
number of distinct states at a block to 10. We could substantially
improve this by tracking only values that affect locking decisions,
which are far fewer than the values that affect all branches.