rtcheck: track all stacks on lock order edges
Currently we just track the first stack we found on a lock order edge.
This is enough to find cycles, but leaves the cycle report wanting.
Change this to track all distinct stack pairs on each lock order edge.
When there's a cycle, this gives a better sense of which order is
"right" if there are significantly more stacks on one order than
another. Also, it shows you everywhere the lock order fails instead of
one arbitrary place, so you don't fix one place only to have the cycle
shift to another stack. This is particularly useful for self cycles.