rtcheck: fix live value checking in block cache
Currently the live value set is only the live values at entry to a
block, which means it never includes phis defined by the current
block. These phi values are provided by the predecessor block and may
participate in future control flow, but since they're not in the live
set, they're ignored when checking the block cache. This means we can
prematurely terminate a path that may lead to different control flow
decisions. For example, this happens in runtime.gcStart and causes it
to not explore the path that calls gcBgMarkStartWorkers.

Fix this by including all live values *in* a block in its live set,
including those that are defined by that block.