rtcheck: fix handling of built-in functions
Currently functions like print and len cause the control flow to stop
because the callee set of the call is empty. That means we're missing
*a lot* of paths.

Fix this by assuming that such functions don't affect the lock state.
This fix explodes the state space, so trim it back down by ignoring
the bodies of throw and traceEvent, and by cutting off a path if it
encounters a self-deadlock.