rtcheck: make HTML much more compact
This uses a much more efficient JSON encoding of the lock graph data,
which reduces the HTML size for the runtime by about an order of a
magnitude.

The main trick is that the strings are interned into small integers.
We also replace the array of structs used for the stack with a struct
of arrays, which reduces redundant field names.