rtcheck: ignore cgoContextPCs and callCgoSymbolizer
These call cgocall or asmcgocall dependent on whether or not we're
panicking or on the system stack. Since we always assume that we don't
know, we assume that they may call cgocall, which takes all sorts of
complicated locks. This appears to be responsible for the majority of
our false positives at the moment, so for now assume they always call
asmcgocall (which is to say, they don't affect the lock graph).