rtcheck: explicit path state representation
Currently walkBlock maintains the path state through a set of local
variables and arguments. The only part of that path state that can be
forked within a block is the lock set. This is going to be a problem
for conditional locks, which will have to fork the value state.

Hence, wrap the entire path state in an explicit type and rewrite
walkBlock in terms of maps and flat-maps across the path state set.