rtcheck: unbind unknown values from the ValState
Currently extending a ValState with an unknown dynamic value is a
no-op. However, it is possible to come back around a loop where some
value was known on an earlier iteration but becomes unknown, and the
current behavior means it will keep its value from the previous
iteration.

Fix this by making Extend with an unknown value "unbind" the value in
the ValState.