rtcheck: model heap and use it to track user vs g0 stack
This adds basic tracking of heap objects to the ValState, and makes
the heap part of the ValState part of the key to the function analysis
memo table.

Using this, it models the result of getg(), the m structure, and
m.curg, so we can track conditions that depend on whether we're on the
user stack or the system stack.

This eliminates a huge number of false positives that appeared when we
started modeling builtin functions (most notably append). This also
prepares us to model m.locks and m.preemptoff, which should eliminate
more false positives; and to model the morestack prologue call, which
should increase coverage.