rtcheck: model morestack prologue
Every function that isn't marked //go:nosplit has an implicit call to
morestack at the beginning. Add this during AST rewriting.

Model morestack as a special function that may always do nothing, and
if we're not on the system stack, may call newstack on the system
stack.

This reveals many more lock edges as well as more feasible deadlocks.
This is particularly useful because these deadlocks are hiding behind
stack splits happening by chance at the wrong time, which is something
user code can generally trigger in the runtime.