rtcheck: fix broken ValState.EqualAt
ValState.EqualAt had a fast path for an empty "at" set that
incorrectly ignored the heap state in addition to the value state. As
a result, we were hitting in the function cache ignoring the heap
state, so we weren't correctly tracking user versus system stack.

Remove this broken fast path. This reveals a huge number of new lock
edges and cycles that were being missed. :(