rtcheck: thread heap state through function calls
Currently the heap state is not threaded through function calls, which
means we can only track heap changes that are "scoped". So far we've
only tracked switches to and from the system stack, for which this is
perfectly adequate, but other things like m.locks don't follow this
pattern.

Fix this by tracking heap state across functions just like we track
lock state.