rtcheck: analyze all runtime roots
This gets the list of runtime entry points from the compiler and uses
all of these functions as analysis roots.

The main difficultly is in generating root calls to these functions
for pointer analysis. To do this, we find each root's declaration in
the AST and use its parsed signature to build a well-formed call to
that function in an init function.