rtcheck: accept ssa.Values in ValState
Currently ValState only accepts ssa.Instructions. I originally did
this because I thought any ssa.Value that wasn't an ssa.Instruction
has to be a constant value. That's not actually true: parameters and
free variables are non-constant ssa.Values that are not
ssa.Instructions.

To make these bindable in a ValState, switch ValState and liveness
over to using ssa.Values and explicitly exclude constant ssa.Values.