rtcheck: track arguments to runtime.copystack
The "sync" argument to runtime.copystack significantly affects its
locking behavior. Currently we don't track argument values, so we
conservatively assume that this argument can take either value, which
leads to a large number of false positives.

Fix this by registering specific functions to track the arguments of.
This keeps the state space reasonable since it requires opting in
specific functions, but lets us handle such cases when they come up.

This substantially reduces false positive edges. There are still a few
that arise from chan.go:recv, which takes an unlock closure. One
possible closure is selunlock, which contains an unlock loop, so we
conservatively assume that the loop may execute zero times and hence
miss the unlock on that path. Since pointer analysis is not
path-sensitive, all calls to recv assume they could call selunlock, so
this happens in the single channel code as well.

There are still a large number of cycles, nearly all of which pass
through _Gscan -> runtime.hchan.lock*. These appear to be legitimate
code paths, though my hunch is that there's a subtle argument for why
this edge is okay.