go-weave: simplify handling of thread aborts
Currently the scheduler is distributed across all of the threads being
run. This certainly simplifies some things, but it makes it much
harder to deal with panics and thread aborts, since they can happen in
many different places. For example, the current implementation freezes
when Sched() is called from a defer run by a thread abort.

Simplify this handling by centralizing the scheduler. All scheduling
decisions are now made directly on the Run goroutine. When a thread
deschedules, it doesn't attempt to pick the next thread to run; it
simply wakes the central scheduler.