gover: fix crash when saving old Go versions
Currently if gover tries to save a directory tree that doesn't exist,
it crashes with a nil pointer dereference because the filepath.Walk
passes an error with a nil info. This happens on older Go versions,
which don't have pkg/include.

Fix this by handling errors in the walk callback.