abi: change analysis to use N registers, more options
Currently, the ABI analysis assumes infinite registers and sees how
many are used. But this doesn't tell you how much stack space is used
for a given number of available registers.

This commit rewrites the ABI analysis to do a sweep of various
configurations (number of registers as well as other ABI options).