stress2: log running and passing commands
This logs running command output to the output directory so users can
check on their progress, and it saves logs of passing commands (not
just failing commands) to the output directory so users can check that
passing commands are working as intended. In both cases, the log files
are saved to hidden files to avoid cluttering the failure logs.