greplogs: add a flag to omit known-broken builders
With a low enough failure rate, it is feasible to triage recent
builder failures using 'greplogs -E . --since=$WHEN'.

However, when turning up a new port the new builder is expected to
fail until the port is complete, and does not require the attention of
anyone who is not actively working on that port. A new flag,
'--omit', allows those builders to be easily filtered out.