greplogs: add a --before flag and parse dates from directory names
It turns out that the dates in the .rev.json files are the repo commit
dates, not the “more recent of {Go date, repo date}” logic that
fetchlogs uses for the directory name. That makes it unsuitable for
filtering, because it classifies new Go-commit runs of old subrepo
commits as old, causing them to be missed.

As an added bonus, parsing the timestamp from the directory name
allows us to avoid opening the .rev.json file entirely, which allows
greplogs to skip nonmatching commits much more efficiently.