greplogs: refine output for issue triage
When --md is set, echo the shell-escaped 'greplogs' command line
itself at the beginning of the output to make the command easier to
reproduce when pasted into an issue report.

Add a --triage flag that adds checkboxes and a summary to the Markdown
output.

Apply the --omit flag to commits as well as builders, to more easily
filter out commits (especially subrepo commits) that are broken across
the board.