minutes3: don't post closed discussions
Currently we don't post locked discussions. There used to be no way to
'close' a discussion so we locked them instead. GitHub eventually
added the ability to close them, and since then we've been using that
instead of locking them. However, minutes3 hadn't caught up with this
change and was still posting closed discussions.