minutes3: close Declined proposals with "not planned" reason
GitHub has added dedicated statuses to help distinguish issues that are
"completed" from ones that are "not planned". When closing a proposal
because it is Declined, the latter status is a better fit.

This should help keep the issue tracker more readable, and, for people
doing issue triage it removes any temptation to manually correct this.

For go.dev/issue/60761.